Use the Websocket-enabled executor
In k8s 1.29, a new v5 remotecommand protocol was released, which uses
Websockets rather than (deprecated) SPDY.

This could give us better chances of stable connections, so let's move
over to it.

We borrow the 'fallback executor' code from `kubectl`, to support both
old and new clusters.

This is enabled on clusters with the
`TranslateStreamCloseWebsocketRequests` flag, which is defaulting to
true since 1.30.