Refactor `Attach` to prevent hanging
We've discovered that we can't reliably wait for the `attacher.Attach`
call to complete, and therefore the CLI to terminate.

To mitigate this, we put the attachment in a Goroutine, and then
_separately_ wait for the completion of the console pod.

Apart from adding this concurrency, it does allow a nice cleanup of the
attach logic, especially given that we don't have to handle the 'detach'
case.